Moved tests to the SF recommended location
